Use Roman numerals directly
While  CCL  claims  to  understand named  Unicode  characters,  in  practice,  it  is  either sorely  lacking,  or  uses
a significantly different notation  than SBCL. In the string-utils case, I spend  some effort transcoding the whitespace
characters and friends  to their CCL Unicode notations, but  in this file, reading Roman numerals  as bizarre codepoints
seemed painful, and maintaining two copies of the tables seemed error-prone, and … whatever. Here it is.